deleted Order entity controller and service

+++ src/main/java/com/example/ajaxonboardingproject/service/mapper/CinemaHallMapper.kt
@@ -17,7 +17,7 @@ class CinemaHallMapper : RequestDtoMapper<CinemaHallRequestDto, CinemaHall>,
 
     override fun mapToDto(model: CinemaHall): CinemaHallResponseDto {
         return CinemaHallResponseDto(
-            id = model.id.toLong(),
+            id = model.id,
             capacity = model.capacity,
             description = model.description
         )
